Small cleanup before RC
