+++ src/App.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import './App.css';
 import TerraSummary from "./terrasummary/terrasummary";
+import {mockData} from "./mockdata";
+
+// Use <TerraSummary data={mockData} /> to use mock data
 
 function App() {
