+++ src/mockdata.js
@@ -1,0 +1,249 @@
+export const mockData = {
+   "1":{
+      "FinalScore":0,
+      "TerraformingRating":21,
+      "ResourceData":{
+         "MegaCredit":{
+            "Quantity":39,
+            "Production":0
+         },
+         "Steel":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Titanium":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Plant":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Energy":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Heat":{
+            "Quantity":0,
+            "Production":3
+         }
+      },
+      "CardsInHand":[
+         "Colonizer Training Camp",
+         "Zeppelins",
+         "Ants",
+         "Martian Rails",
+         "Asteroid Mining Consortium",
+         "Shuttles",
+         "Imported Nitrogen",
+         "Trans-Neptune Probe",
+         "Nitrite Reducing Bacteria",
+         "Energy Tapping",
+         "Capital",
+         "Mangrove",
+         "Urbanized Area",
+         "Ice cap Melting",
+         "Industrial Microbes",
+         "Building Industries"
+      ],
+      "CardsPlayed":[
+         "Subterranean Reservoir"
+      ],
+      "PlayerId":1,
+      "IsLocalPlayer":true,
+      "Color":"Green",
+      "Score":{
+         "FinalScore":43,
+         "TerraFormingRating":21,
+         "VictoryPoints":0,
+         "GreeneryScore":0,
+         "CityScore":15,
+         "MilestoneScore":[
+            {
+               "Name":"Planner",
+               "Score":5
+            }
+         ],
+         "AwardScore":[
+            {
+               "Name":"Landlord",
+               "Score":2
+            },
+            {
+               "Name":"Banker",
+               "Score":0
+            },
+            {
+               "Name":"Scientist",
+               "Score":0
+            },
+            {
+               "Name":"Thermalist",
+               "Score":0
+            },
+            {
+               "Name":"Miner",
+               "Score":0
+            }
+         ]
+      }
+   },
+   "2":{
+      "FinalScore":0,
+      "TerraformingRating":26,
+      "ResourceData":{
+         "MegaCredit":{
+            "Quantity":9,
+            "Production":0
+         },
+         "Steel":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Titanium":{
+            "Quantity":2,
+            "Production":0
+         },
+         "Plant":{
+            "Quantity":1,
+            "Production":1
+         },
+         "Energy":{
+            "Quantity":0,
+            "Production":1
+         },
+         "Heat":{
+            "Quantity":4,
+            "Production":2
+         }
+      },
+      "CardsInHand":[
+         "Nitrophilic Moss",
+         "Fish",
+         "Physics Complex"
+      ],
+      "CardsPlayed":[
+         "Technology Demonstration",
+         "Dust Seals",
+         "Geothermal Power",
+         "Development Center",
+         "Deimos Down",
+         "Tardigrades",
+         "Magnetic Field Dome",
+         "Power Plant"
+      ],
+      "PlayerId":2,
+      "IsLocalPlayer":false,
+      "Color":"Grey",
+      "Score":{
+         "FinalScore":43,
+         "TerraFormingRating":26,
+         "VictoryPoints":1,
+         "GreeneryScore":10,
+         "CityScore":0,
+         "MilestoneScore":[
+
+         ],
+         "AwardScore":[
+            {
+               "Name":"Landlord",
+               "Score":2
+            },
+            {
+               "Name":"Banker",
+               "Score":0
+            },
+            {
+               "Name":"Scientist",
+               "Score":2
+            },
+            {
+               "Name":"Thermalist",
+               "Score":2
+            },
+            {
+               "Name":"Miner",
+               "Score":0
+            }
+         ]
+      }
+   },
+   "3":{
+      "FinalScore":0,
+      "TerraformingRating":21,
+      "ResourceData":{
+         "MegaCredit":{
+            "Quantity":10,
+            "Production":1
+         },
+         "Steel":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Titanium":{
+            "Quantity":0,
+            "Production":0
+         },
+         "Plant":{
+            "Quantity":2,
+            "Production":1
+         },
+         "Energy":{
+            "Quantity":0,
+            "Production":3
+         },
+         "Heat":{
+            "Quantity":0,
+            "Production":0
+         }
+      },
+      "CardsInHand":[
+         "Commercial District",
+         "Aerobraked Ammonia Asteroid"
+      ],
+      "CardsPlayed":[
+         "Invention Contest",
+         "Lagrange Observatory",
+         "Search For Life",
+         "Archaebacteria",
+         "Earth Office",
+         "Regolith Eaters",
+         "Tectonic Stress Power"
+      ],
+      "PlayerId":3,
+      "IsLocalPlayer":false,
+      "Color":"Yellow",
+      "Score":{
+         "FinalScore":40,
+         "TerraFormingRating":21,
+         "VictoryPoints":2,
+         "GreeneryScore":1,
+         "CityScore":1,
+         "MilestoneScore":[
+
+         ],
+         "AwardScore":[
+            {
+               "Name":"Landlord",
+               "Score":5
+            },
+            {
+               "Name":"Banker",
+               "Score":0
+            },
+            {
+               "Name":"Scientist",
+               "Score":5
+            },
+            {
+               "Name":"Thermalist",
+               "Score":5
+            },
+            {
+               "Name":"Miner",
+               "Score":0
+            }
+         ]
+      }
+   }
+};
